homelane/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-cors-headers==3.11.0
django-sslserver==0.22
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
idna==3.3
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
